filter_samples.py
